+++ examples/object_storage/bucket.tf
@@ -20,27 +20,6 @@ resource "oci_objectstorage_bucket" "bucket_with_versioning" {
   versioning     = "Suspended"
 }
 
-resource "oci_objectstorage_object_lifecycle_policy" "lifecyclePolicy1" {
-  namespace = "${data.oci_objectstorage_namespace.ns.namespace}"
-  bucket    = "${oci_objectstorage_bucket.bucket1.name}"
-
-  #Optional
-  rules {
-    #Required
-    action      = "ARCHIVE"
-    is_enabled  = "true"
-    name        = "test-rule-1"
-    time_amount = "10"
-    time_unit   = "DAYS"
-
-    #Optional
-    object_name_filter {
-      #Optional
-      inclusion_prefixes = ["my-test"]
-    }
-  }
-}
-
 data "oci_objectstorage_bucket_summaries" "buckets1" {
   compartment_id = "${var.compartment_ocid}"
   namespace      = "${data.oci_objectstorage_namespace.ns.namespace}"
@@ -53,17 +32,6 @@ data "oci_objectstorage_bucket_summaries" "buckets1" {
 
 output buckets {
   value = "${data.oci_objectstorage_bucket_summaries.buckets1.bucket_summaries}"
-}
-
-data "oci_objectstorage_object_lifecycle_policy" "lifecyclePolicies1" {
-  namespace = "${data.oci_objectstorage_namespace.ns.namespace}"
-  bucket    = "${oci_objectstorage_bucket.bucket1.name}"
-
-  depends_on = ["oci_objectstorage_object_lifecycle_policy.lifecyclePolicy1"]
-}
-
-output lifecyclePolicies1 {
-  value = "${data.oci_objectstorage_object_lifecycle_policy.lifecyclePolicies1.rules}"
 }
 
 data "oci_objectstorage_object_versions" "test_object_versions1" {