+++ oci/objectstorage_object_lifecycle_policy_resource.go
@@ -127,6 +127,11 @@ func ObjectStorageObjectLifecyclePolicyResource() *schema.Resource {
 								},
 							},
 						},
+						"target": {
+							Type:     schema.TypeString,
+							Optional: true,
+							Default:  "objects",
+						},
 
 						// Computed
 					},
@@ -392,6 +397,11 @@ func (s *ObjectStorageObjectLifecyclePolicyResourceCrud) mapToObjectLifecycleRul
 		}
 	}
 
+	if target, ok := s.D.GetOkExists(fmt.Sprintf(fieldKeyFormat, "target")); ok {
+		tmp := target.(string)
+		result.Target = &tmp
+	}
+
 	if timeAmount, ok := s.D.GetOkExists(fmt.Sprintf(fieldKeyFormat, "time_amount")); ok {
 		tmp := timeAmount.(string)
 		tmpInt64, err := strconv.ParseInt(tmp, 10, 64)
@@ -425,6 +435,10 @@ func ObjectLifecycleRuleToMap(obj oci_object_storage.ObjectLifecycleRule) map[st
 
 	if obj.ObjectNameFilter != nil {
 		result["object_name_filter"] = []interface{}{ObjectNameFilterToMap(obj.ObjectNameFilter)}
+	}
+
+	if obj.Target != nil {
+		result["target"] = string(*obj.Target)
 	}
 
 	if obj.TimeAmount != nil {
@@ -547,6 +561,9 @@ func rulesHashCodeForSets(v interface{}) int {
 				}
 			}
 		}
+	}
+	if target, ok := m["target"]; ok && target != "" {
+		buf.WriteString(fmt.Sprintf("%v-", target))
 	}
 	if timeAmount, ok := m["time_amount"]; ok {
 		buf.WriteString(fmt.Sprintf("%v-", timeAmount))