+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## 3.89.0 (Unreleased)
+### Added
+- Object Lifecycle Management support for Multipart Uploads Cleanup
+
 ## 3.88.0 (August 05, 2020)
 
 ### Added